# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Referral & Commission Platform API
  version: 1.0.0-draft.1
  summary: Phase 0 draft of the v1 API (spec §20). Generated docs replace this file from Phase 2 on; until then this file is the contract.
  description: |
    Conventions (see docs/api.md):
    - All paths are under `/v1`; the gateway may expose them as `/api/v1`.
    - Money is integer minor units plus ISO-4217 currency. Never floats.
    - Timestamps are RFC 3339 in UTC.
    - List endpoints use cursor pagination: `?cursor=&limit=` (max 100), response `{items, next_cursor}`.
    - Successful responses return the resource directly; errors use the `Error` envelope.
    - `Idempotency-Key` is required on every non-event POST that moves money.
    - Service endpoints act only for the product bound to the API credential. A product is never taken from the request body.
  license:
    name: Proprietary
servers:
  - url: https://api.ref.example.com/v1
  - url: http://localhost:8000/v1
    description: local
...
security: []
paths:
...
  /r/{code}:
    get:
      tags: [public]
      operationId: referralRedirect
      summary: Record a click and redirect to the product's registered signup URL (§7.1)
      description: |
        Validates the code, records a click (hashed IP, UA, UTM, campaign), issues a signed attribution token,
        sets a first-party cookie on the referral domain, and redirects to the product's registered `signup_url`
        with `?ref={CODE}&rt={token}`. Redirect targets come only from the product registry.
        Without `product`, renders a landing page listing products the code works for.
        An invalid or inactive code renders a neutral landing page (no enumeration signal beyond `/public/codes`).
      parameters:
        - $ref: '#/components/parameters/ReferralCodePath'
        - name: product
          in: query
          schema: { $ref: '#/components/schemas/ProductSlug' }
        - name: campaign
          in: query
          schema: { type: string, maxLength: 64 }
        - name: utm_source
          in: query
          schema: { type: string, maxLength: 128 }
        - name: utm_medium
          in: query
          schema: { type: string, maxLength: 128 }
        - name: utm_campaign
          in: query
          schema: { type: string, maxLength: 128 }
      responses:
        '302':
          description: Redirect to the registered signup URL
          headers:
            Location: { schema: { type: string, format: uri } }
            Set-Cookie: { schema: { type: string } }
        '200':
          description: Landing page (HTML) when `product` is missing or the code is not valid for it
          content:
            text/html: { schema: { type: string } }
        '429': { $ref: '#/components/responses/RateLimited' }
  /public/codes/{code}:
    get:
      tags: [public]
      operationId: checkReferralCode
      summary: Check whether a referral code is valid (rate-limited per IP, §6)
      parameters:
        - $ref: '#/components/parameters/ReferralCodePath'
      responses:
        '200':
          description: Result. Unknown codes return `valid=false` with 200, not 404.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicCodeCheck' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /referrers/enroll:
    post:
      tags: [service]
      operationId: enrollReferrer
      summary: Enroll the product's logged-in user as a referrer (§5.2). Idempotent per product account.
      security: [ { serviceKey: [ 'referrers:enroll' ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollRequest' }
      responses:
        '201':
          description: Enrolled
          content: { application/json: { schema: { $ref: '#/components/schemas/EnrollResponse' } } }
        '200':
          description: Already enrolled
          content: { application/json: { schema: { $ref: '#/components/schemas/EnrollResponse' } } }
        '422': { $ref: '#/components/responses/Error' }
  /attributions:
    post:
      tags: [service]
      operationId: createAttribution
      summary: Attribute a newly registered product account to a referrer (§7.2–7.3)
      description: |
        Idempotent on (product, external_user_id): 201 new, 200 existing (returns the existing attribution, even if
        a different code is sent). Rejections return 422 with a reason code. Product SDKs fail soft: an error here
        must never block signup.
      security: [ { serviceKey: [ 'attribution:write' ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttributionRequest' }
      responses:
        '201':
          description: Attributed
          content: { application/json: { schema: { $ref: '#/components/schemas/Attribution' } } }
        '200':
          description: Already attributed (existing attribution returned)
          content: { application/json: { schema: { $ref: '#/components/schemas/Attribution' } } }
        '422':
          description: |
            Rejected. Codes: REFERRAL_CODE_INVALID, ATTRIBUTION_TOKEN_INVALID, ATTRIBUTION_TOKEN_EXPIRED,
            ATTRIBUTION_NOT_NEW_ACCOUNT, ATTRIBUTION_TIMESTAMP_INVALID, SELF_REFERRAL, REFERRER_BLOCKED,
            REFERRAL_CODE_REQUIRED.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  /reconciliation/payments:
    post:
      tags: [service]
      operationId: reconcilePayments
      summary: Daily payment list for referred accounts to detect missed events (ADR 0017, Phase 13)
      security: [ { serviceKey: [ 'events:write' ] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [business_date, payments]
              properties:
                business_date: { type: string, format: date }
                payments:
                  type: array
                  maxItems: 10000
                  items:
                    type: object
                    additionalProperties: false
                    required: [payment_id, external_user_id, amount_net_paid, currency, paid_at]
                    properties:
                      payment_id: { $ref: '#/components/schemas/ExternalId' }
                      external_user_id: { $ref: '#/components/schemas/ExternalId' }
                      amount_net_paid: { $ref: '#/components/schemas/MinorAmount' }
                      currency: { $ref: '#/components/schemas/Currency' }
                      paid_at: { $ref: '#/components/schemas/Timestamp' }
      responses:
        '200':
          description: Reconciliation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id: { type: string, format: uuid }
                  missing_on_platform: { type: array, items: { $ref: '#/components/schemas/ExternalId' } }
                  amount_mismatch: { type: array, items: { $ref: '#/components/schemas/ExternalId' } }

  # ---------------------------------------------------------------- me
  /me:
    get:
      tags: [me]
      operationId: getMe
      security: [ { userJwt: [] } ]
      responses:
        '200':
          description: Current user
          content: { application/json: { schema: { $ref: '#/components/schemas/Me' } } }
...
components:
  securitySchemes:
    serviceKey:
      type: http
      scheme: bearer
      bearerFormat: '<key_id>.<secret>'
      description: Product API client credential with scopes. Never sent from a browser.
    eventSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 over `timestamp + "." + raw_body` with the client's signing secret.
    userJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 15-minute access token held in memory by the portal/admin.
  parameters:
    ReferralCodePath:
      name: code
      in: path
      required: true
      schema: { $ref: '#/components/schemas/ReferralCodeValue' }
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string, pattern: '^[A-Z][A-Z0-9_]+$', description: See the error registry in docs/api.md }
            message: { type: string }
            request_id: { type: string }
            details: { type: object }
    Timestamp: { type: string, format: date-time }
    Currency: { type: string, pattern: '^[A-Z]{3}$' }
    MinorAmount: { type: integer, minimum: 0, format: int64 }
...
    ExternalId: { type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$' }
    ProductSlug: { type: string, pattern: '^[a-z][a-z0-9-]{1,31}$' }
...
    ReferralCodeValue:
      type: string
      description: 4–16 chars from A–Z and 2–9 excluding 0 O 1 I L. Case-insensitive on input.
      pattern: '^[A-HJKMNP-Za-hjkmnp-z2-9]{4,16}$'
    Signals:
      type: object
      additionalProperties: false
      description: Raw signals; HMAC-hashed on receipt, never stored or logged raw.
      properties:
        ip: { type: string, maxLength: 45 }
        email: { type: string, maxLength: 254 }
        phone: { type: string, maxLength: 20 }
        device_id: { type: string, maxLength: 128 }
    PublicCodeCheck:
      type: object
      required: [valid]
      properties:
        valid: { type: boolean }
        referrer_display_name: { type: [string, 'null'], description: Only when valid }
...
    EnrollRequest:
      type: object
      additionalProperties: false
      required: [external_user_id]
      properties:
        external_user_id: { $ref: '#/components/schemas/ExternalId' }
        display_name: { type: string, maxLength: 80 }
        email: { type: string, format: email, description: Used for the platform user profile only; never used to link accounts (§5.1) }
        phone: { type: string, maxLength: 20 }
        locale: { type: string, enum: [en, bn] }
        signals: { $ref: '#/components/schemas/Signals' }
    EnrollResponse:
      type: object
      required: [user_id, code, portal_url, link]
      properties:
        user_id: { type: string, format: uuid }
        code: { $ref: '#/components/schemas/ReferralCodeValue' }
        link: { type: string, format: uri, description: 'https://ref.example.com/r/{CODE}?product={slug}' }
        portal_url: { type: string, format: uri }
    AttributionRequest:
      type: object
      additionalProperties: false
      required: [external_user_id, registered_at]
      anyOf:
        - required: [referral_code]
        - required: [attribution_token]
      properties:
        external_user_id: { $ref: '#/components/schemas/ExternalId' }
        registered_at: { $ref: '#/components/schemas/Timestamp' }
        referral_code: { $ref: '#/components/schemas/ReferralCodeValue' }
        attribution_token: { type: string, maxLength: 2048 }
        signals: { $ref: '#/components/schemas/Signals' }
    Attribution:
      type: object
      required: [referral_id, external_user_id, status, source, attributed_at]
      properties:
        referral_id: { type: string, format: uuid }
        external_user_id: { $ref: '#/components/schemas/ExternalId' }
        referral_code: { $ref: '#/components/schemas/ReferralCodeValue' }
        campaign: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/ReferralStatus' }
        source: { type: string, enum: [TYPED_CODE, TOKEN, TYPED_AND_TOKEN, ADMIN] }
        attributed_at: { $ref: '#/components/schemas/Timestamp' }
    ReferralStatus:
      type: string
      enum: [ATTRIBUTED, TRIALING, CONVERTED, ACTIVE, INACTIVE, EXPIRED, REJECTED, REVOKED]
...
    Me:
      type: object
      required: [id, roles]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        email: { type: [string, 'null'] }
        email_verified: { type: boolean }
        phone_verified: { type: boolean }
        kyc_status: { type: string, enum: [NONE, PENDING, VERIFIED, REJECTED] }
        roles: { type: array, items: { type: string } }
        locale: { type: string, enum: [en, bn] }
        time_zone: { type: string }
...